Refactor project into feature folders
