+++ ReactNative/Great_Vinyl_Exchange/.eslintrc.json
@@ -1,3 +1,19 @@
 {
-    "extends": "rallycoding"
+    "parser": "babel-eslint",
+    "env": {
+        "browser": true
+    },
+    "plugins": [
+        "react",
+        "import"
+    ],
+    "extends": [
+        "eslint:recommended",
+        "plugin:react/recommended",
+        "plugin:import/errors",
+        "plugin:import/warnings"
+    ],
+    "rules": {
+        // overrides
+    }
 }
