+++ ReactNative/Great_Vinyl_Exchange/package.json
@@ -8,6 +8,7 @@
 	},
 	"dependencies": {
 		"firebase": "^4.6.0",
+		"keymirror": "^0.1.1",
 		"lodash": "^4.17.4",
 		"prop-types": "^15.5.10",
 		"react": "16.0.0-alpha.12",
