+++ ReactNative/Great_Vinyl_Exchange/src/ReducerIndex.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import AuthenticationReducer from './AuthenticationReducer';
-import NavigationReducer from './NavigationReducer';
+import AuthenticationReducer from './authentication/AuthenticationReducer';
+import NavigationReducer from './navigation/NavigationReducer';
 
 export default combineReducers({
     authentication: AuthenticationReducer,
