--- ReactNative/Great_Vinyl_Exchange/src/actions/index.js
@@ -1,1 +1,0 @@
-export * from './AuthenticationActions';
