--- ReactNative/Great_Vinyl_Exchange/src/actions/types.js
@@ -1,12 +1,0 @@
-// Authentication
-export const SET_EMAIL = "set_email";
-export const SET_USERNAME = "set_username";
-export const SET_PASSWORD = "set_password";
-export const SET_ADDRESS = "set_address";
-export const REGISTER_USER = "register_user";
-export const REGISTER_USER_SUCCESS = "register_user_success";
-export const REGISTER_USER_FAIL = "register_user_fail";
-export const ADD_USER_SUCCESS = 'add_user_success';
-
-// Navigation
-export const CURRENT_SCENE = "current_scene";
