+++ ReactNative/Great_Vinyl_Exchange/src/app/AppRoot.js
@@ -1,10 +1,9 @@
 import React, { Component } from 'react';
 import firebase from 'firebase';
 import { Provider } from 'react-redux';
-import { View } from 'react-native';
 import ReduxThunk from 'redux-thunk';
 import { createStore, applyMiddleware } from 'redux';
-import reducers from '../reducers';
+import reducers from '../ReducerIndex';
 import AppNavigator from '../navigation/AppNavigator';
 
 //import getStore from '../store/store';
