+++ ReactNative/Great_Vinyl_Exchange/src/authentication/AuthenticationActions.js
@@ -1,49 +1,50 @@
 import firebase from 'firebase';
 import { NavigationActions } from 'react-navigation';
 
-import { SET_EMAIL,
-         SET_USERNAME,
-         SET_PASSWORD, 
-         SET_ADDRESS, 
-         REGISTER_USER,
-         REGISTER_USER_SUCCESS, 
-         REGISTER_USER_FAIL,
-         ADD_USER_SUCCESS,
-} from './types';
+const { 
+    AUTHENTICATION_SET_EMAIL,
+    AUTHENTICATION_SET_USERNAME,
+    AUTHENTICATION_SET_PASSWORD, 
+    AUTHENTICATION_SET_ADDRESS, 
+    AUTHENTICATION_REGISTER_USER,
+    AUTHENTICATION_REGISTER_USER_SUCCESS, 
+    AUTHENTICATION_REGISTER_USER_FAIL,
+    AUTHENTICATION_ADD_USER_SUCCESS,
+} = require('../resources/ActionConstants').default;
 
 export const setEmail = (email) => {
     return {
-        type: SET_EMAIL,
+        type: AUTHENTICATION_SET_EMAIL,
         payload: email
     };
 };
 
 export const setUsername = (username) => {
     return {
-        type: SET_USERNAME,
+        type: AUTHENTICATION_SET_USERNAME,
         payload: username
     };
 };
 
 export const setPassword = (password) => {
     return {
-        type: SET_PASSWORD,
+        type: AUTHENTICATION_SET_PASSWORD,
         payload: password
     };
 };
 
 export const setAddress = (address) => {
     return {
-        type: SET_ADDRESS,
+        type: AUTHENTICATION_SET_ADDRESS,
         payload: address
     };
 };
 
 export const registerUser = (user, password) => {
     return (dispatch) => {
-        dispatch({type: REGISTER_USER });
+        dispatch({type: AUTHENTICATION_REGISTER_USER });
         firebase.auth().createUserWithEmailAndPassword(user.email, password)
-            .then(firebaseUser => registerUserSuccess(dispatch, user))
+            .then(() => registerUserSuccess(dispatch, user))
             .catch(function(error) {
                 registerUserFail(dispatch, error);
             });
@@ -52,16 +53,15 @@ export const registerUser = (user, password) => {
 
 const registerUserFail = (dispatch, error) => {
     dispatch({ 
-        type: REGISTER_USER_FAIL,
+        type: AUTHENTICATION_REGISTER_USER_FAIL,
         payload: error
      })
 };
 
 const registerUserSuccess = (dispatch, user) => {
-    console.log(`sucess ${user} ${user.email}`)
     addUserToDatabase(dispatch, user);
     dispatch({
-        type: REGISTER_USER_SUCCESS,
+        type: AUTHENTICATION_REGISTER_USER_SUCCESS,
         payload: user
     });
 };
@@ -71,7 +71,7 @@ const addUserToDatabase = (dispatch, user) => {
     firebase.database().ref(`/users/${currentUser.uid}/userProfile`)
         .push(user)
         .then(() => {
-            dispatch({ type: ADD_USER_SUCCESS, payload: user });
+            dispatch({ type: AUTHENTICATION_ADD_USER_SUCCESS, payload: user });
             dispatch(NavigationActions.navigate({routeName: 'AuthenticationMain'}));
         })
         .catch(error => {
