+++ ReactNative/Great_Vinyl_Exchange/src/common-components/LinkButton.js
