+++ ReactNative/Great_Vinyl_Exchange/src/common-components/index.js
