+++ ReactNative/Great_Vinyl_Exchange/src/navigation/NavigationReducer.js
