+++ ReactNative/Great_Vinyl_Exchange/src/navigation/navigation.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { StackNavigator } from 'react-navigation';
 
-import AuthenticationMain from '../containers/authentication/AuthenticationMain';
-import AuthenticationSignUpEmail from '../containers/authentication/AuthenticationSignUpEmail';
-import AuthenticationSignUpUsername from '../containers/authentication/AuthenticationSignUpUsername';
-import AuthenticationSignUpAddress from '../containers/authentication/AuthenticationSignUpAddress';
-import AuthenticationSignUpPassword from '../containers/authentication/AuthenticationSignUpPassword';
-import AuthenticationSignUpConfirmPassword from '../containers/authentication/AuthenticationSignUpConfirmPassword';
+import AuthenticationMain from '../authentication/screens/AuthenticationMain';
+import AuthenticationSignUpEmail from '../authentication/screens/AuthenticationSignUpEmail';
+import AuthenticationSignUpUsername from '../authentication/screens/AuthenticationSignUpUsername';
+import AuthenticationSignUpAddress from '../authentication/screens/AuthenticationSignUpAddress';
+import AuthenticationSignUpPassword from '../authentication/screens/AuthenticationSignUpPassword';
+import AuthenticationSignUpConfirmPassword from '../authentication/screens/AuthenticationSignUpConfirmPassword';
 
 export default StackNavigator({
     AuthenticationMain: { screen: AuthenticationMain, navigationOptions: { header: null } },
