+++ ReactNative/Great_Vinyl_Exchange/src/resources/images/back-arrow.png
