+++ ReactNative/Great_Vinyl_Exchange/src/resources/images/gve_logo.png
